DO NOT MERGE: PROVISORY: profiles: qca_ipq95xx: io-expander-freedom: enable power on M.2 connector
FULL_CARD_POWER_OFF# on Quectel RM520N-GL is internally pulled down with
a 100k resistor, so the module is powered off by default.

So enable GPIO 55 which seems to be connected on Freedom board to this
FULL_CARD_POWER_OFF# pin in order to provide power to Quectel RM520N-GL
module, right after the boot.

References: PCF-1173
Signed-off-by: Petr Štetiar <petr.stetiar@prplfoundation.org>