prpl-configuration: split common and target specific files
In the past we have only build for one target (freedom). Since we now want to add mxl-osp as well, we need to split up the prpl-configuration package.
For every box one can enable the package belonging to that specific box:
- prpl-configuration-qca-ipq95xx
- prpl-configuration-mxl-x86-osp-tb341

This will ensure the common files + the specific files for this hardware are installed.

The networklayout.json file specificly needed for wnc-freedom is moved to the freedom specific package. The other files are all common.

Signed-off-by: Matthias FRANCK <matthias.franck_ext@softathome.com>