PROVISORY: .testbed: cdrouter: create separate wifi7 configuration
In 602f983ef7470740c6a09d247d32407a0a2717b7 we updated wifiBeaconPhy to expect the wifi7 beacon.
On boxes without wifi7, the test is now failing (mxlosp). Let's create a separate cdrouter configuration for wifi7 boxes.

A cleaner solution should be found, so that we do not have to duplicate whole CDRouter congirations. This is tracked in PPW-25.

Signed-off-by: Matthias FRANCK <matthias.franck_ext@softathome.com>