ci: cram: 025-process-monitor.t: restore correct CurrentTestInterval
CurrentTestInterval was being overwritten during the test blindly.
Save the CurrentTestInterval value so it can be restored at the end of
the test.

Also process check time can vary depending on the test count. This might
become a problem if test counts increase. Instead of waiting for a fixed
timeout, let's calculate reactivation_timeout with current test count
and use that as a timeout for process restart.

Signed-off-by: Yüce Kürüm <yuce.kurum@mind.be>
Signed-off-by: Yüce Kürüm <yuce.kurum_ext@softathome.com>