ci: cram: add test for tr181-dnssd
There are no tests for checking the DNS Service Discovery functionality.
In order to properly test tr181-dnssd we need to extract the services
from the responses.

`nmap -sU --script=dns-service-discovery -p5353 192.168.1.1` is supposed
to that, but advertised services don't show up in the command output.
We can still see them in the packet capture:
14:40:31.631478 IP 192.168.1.2.53064 > 192.168.1.1.5353: 0 PTR (QM)? _services._dns-sd._udp.local. (46)
14:40:31.634862 IP 192.168.1.1.5353 > 224.0.0.251.5353: 0*- [0q] 1/0/0 A 192.168.1.1 (44)
14:40:31.634862 IP 192.168.1.1.5353 > 224.0.0.251.5353: 0*- [0q] 1/0/0 PTR _mqtt._tcp.local. (70)
14:40:31.634862 IP 192.168.1.1.5353 > 224.0.0.251.5353: 0*- [0q] 1/0/0 PTR prplOS.lan._mqtt._tcp.local. (69)
14:40:31.634862 IP 192.168.1.1.5353 > 224.0.0.251.5353: 0*- [0q] 2/0/0 SRV prplOS.lan.local.:0 1648 29296, TXT "From=sah" (123)

Until we can find a way to properly check the advertised services, let's
at least make sure that the basic functionality is there.

For this purpose:
- Check DM entry.
- Check firewall rules
- Check if service responds

Signed-off-by: Yüce Kürüm <yuce.kurum@mind.be>
Signed-off-by: Yüce Kürüm <yuce.kurum_ext@softathome.com>