profiles: prpl: fix LXC build by explicitly installing packages
In commit 12c08a74add2 ("packages: import LXC v5.0.3 from packages
feed"), we imported the LXC package from the feeds. Since LXC depends on
libseccomp, which remained in the packages feed, this caused a build
failure due to the feeds' interworkings. The libseccomp package is no
longer pulled in automatically, resulting in the following error:

  * opkg_install_cmd: Cannot install package liblxc.
  * satisfy_dependencies_for: Cannot satisfy the following dependencies for lxc-auto:
  *     libseccomp

  * pkg_hash_check_unresolved: cannot find dependency shadow-utils for lxc-unprivileged
  * pkg_hash_check_unresolved: cannot find dependency shadow-newuidmap for lxc-unprivileged
  * pkg_hash_check_unresolved: cannot find dependency shadow-newgidmap for lxc-unprivileged

To fix this, we will explicitly install the libseccomp, shadow-utils,
shadow-newuidmap, shadow-newgidmap package.

References: PCF-1315
Signed-off-by: Petr Štetiar <petr.stetiar@prplfoundation.org>