.gitlab: testbed: run cram tests manually
Every time we make a build, the cram tests are executed automatically. Because of this, tests are often excecuted while not needed and test capacity is occupied for nothing.
Let's adapt this so that we manually need to trigger the test jobs if we would like to excecute cram tests.

Signed-off-by: Matthias FRANCK <matthias.franck@softathome.com>