ci: cram: 095-syslog.t: adapt the test for changed host description
Syslog tests started to fail and it seems like it can't match the log
string when we try to log from a network source.

Network source uses the same template as the system logs.
Syslog.Template.1.Expression="\"$\{FULLDATE\} $\{HOST\} $\{MSGHDR\}$\{MSG\}\\n\""

Seems like logs are printing the IP address instead of the hostname now.
Let's adapt the test so it accepts an IP address instead of a hostname.

Signed-off-by: Yüce Kürüm <yuce.kurum@mind.be>
Signed-off-by: Yüce Kürüm <yuce.kurum_ext@softathome.com>