PROVISORY: profiles: qca_ipq95xx: enable prpl-configuration-qca-ipq95xx
It looks like the order of WiFi.Radio.X interfaces is not deterministic.
This is caused because the networklayout.json file that is generated has
no fixed order of the radio interfaces.  As a workaround, a static
networklayout.json is added for the wnc-freedom box (located in the
prpl-configuration-qca-ipq95xx package).

This issue is related to PPW-20 and should probably be solved by
integrating feed-qca@84bd94fa

References: PPW-20
Signed-off-by: Matthias FRANCK <matthias.franck@softathome.com>