ci: 020-security: add test for Security.Certificate.{i}.Enable
According to the TR-181 specification, it should be possible to disable a
security certificate via Security.Certificate.{i}.Enable=0. However, due to
the field being marked as read-only in the datamodel (%read-only bool
Enable = true), this was not possible. This caused an error when
attempting to set Security.Certificate.{i}.Enable=0.

The issue was fixed with tr181-security version v0.4.2 in commit
8f9f655978c5 ("Issue: HOP-6242 TR-181 Device.Security data model issues
19.03.2024").

So lets add a test case that verifies that certificate with the
SignatureAlgorithm set to ecdsa-with-SHA512 can be disabled and enabled
again.

References: PCF-1054
Signed-off-by: Petr Štetiar <petr.stetiar@prplfoundation.org>