ci: cram: check-for-crashes.t: check ProcessFaults entries as well
It seems that on some platforms/kernels and in some cases the crash is
not simply available in the logs, thus the grep is not going to catch
it, so lets add ProcessFaults. datamodel check as it seems, that in some
cases it can provide hint about some crashes.

Signed-off-by: Petr Štetiar <petr.stetiar@prplfoundation.org>
(cherry picked from commit 2f2d22b93b411ce87d2503c80a28bbe2c3a6d859)