profiles: prpl: take over prpl.yml for M3.2 preparations
In preparation of creating the mainline-3.2 let's take over the changes
in prpl.yml from prplos-23.05. This file gets updated regularly and
causes a lot of merge conflicts.

A detailed git history of the file can be found in prplos-23.05 branch

This version does not use the proposed tag@hash schema for the feed
revisions yet. It will be updated once the CI gets updated.

Signed-off-by: Yüce Kürüm <yuce.kurum@mind.be>
Signed-off-by: Yüce Kürüm <yuce.kurum_ext@softathome.com>