profiles: prpl.yml: lxc patch for problem described in PPW-60
During the setup/start of an LXC container, the LXC creates multiple
forks of the main process for different purposes. Each fork inherits the
file descriptors (fds) of the parent process.  However, each child
process closes the list of inherited file descriptors that it does not
need, as these fds are irrelevant to the child process.  This cleanup of
fds occurs in the lxc_check_inherited().

If logging to syslog is enabled, the socket used for logging to the
syslog daemon (through the syslog device: /dev/log) is also closed,
invalidating the corresponding file descriptor.  LXC uses the standard
syslog() function to log messages to the syslog daemon.

OpenWrt uses the musl library as its standard C library, and the musl
implementation of syslog() is not resilient to the invalidation of
previously obtained file descriptors (through calls to syslog() or
openlog()).  In contrast, the syslog() implementation in glibc and other
C libraries is more robust: if an invalid file descriptor is detected,
those libraries will automatically reopen the connection to the syslog
device, obtaining a new, valid file descriptor to continue logging.

musl libc, on the other hand, only detects connection loss to the syslog
device and does not handle cases where the file descriptor itself
becomes invalid.  This limitation can be seen in the musl implementation
here: https://git.musl-libc.org/cgit/musl/tree/src/misc/syslog.c#n80.
musl libc behavior will not change following the answer maintainers:
https://www.openwall.com/lists/musl/2024/09/26/2

As a result of this behavior in musl, when an LXC child process tries to
log information using syslog(), it continues to use the invalid file
descriptor (which was closed earlier).  This primarily results in some
missing logs without functional impacts, when no other socket or file
opening operation reuses the same file descriptor value. However, if the
file descriptor is reused by another process or operation, logs may
erroneously be written to that file or socket and can lead to container
creation/startup failure.

This commit aims to prevent forked processes from using invalid syslog
file descriptors, which can lead to issues such as missing logs or, in
worse cases, random failures in LXC functionality (e.g., when
synchronization between LXC processes is disrupted).  It also addresses
musl's lack of resilience when handling invalid file descriptors.

The changes work by caching the syslog connection information at the
start and reusing it to reopen the syslog connection after child
processes close all inherited file descriptors.

References: PPW-60
Tested-by: Petr Štetiar <petr.stetiar@prplfoundation.org>
Implemented-by: Nabil BIZID <nabil.bizid@softathome.com>
Signed-off-by: Yüce Kürüm <yuce.kurum@mind.be>
Signed-off-by: Yüce Kürüm <yuce.kurum_ext@softathome.com>