ci: cram: Add libacl config option
In feed_lcm v4.0.1, a dependency on libacl is added to add the functionality to have containers run with an unprivileged user.
A new user will be created for each container, and the ownership of the
layers will either be modified through idmap mounting the layers if
available, or by modifying the acl of the layers with posix acl.

Issue: LCM-740 - Support for unprivileged containers
Signed-off-by: Toon Caeyers <toon.caeyers_ext@softathome.com>
(cherry picked from commit f8163ea7643995807ca1b16e52c9aec4e0062394)