profiles: prpl: add PCI/USB/serial port utils to ease work with cellular modems
Currently, the system lacks tools to inspect USB and PCI buses or access
serial ports, which makes working with cellular modems inconvenient.
This commit addresses the issue by adding the `pci-utils`, `usb-utils`,
and `picocom` packages. These tools are not only useful for development
but are also planned for use in CI QA workflows.

Since the `debug.yml` profile is not utilized during CI, the tools are
added directly to the prpl.yml profile instead of the debug profile.

References: PCF-1173
Signed-off-by: Petr Štetiar <petr.stetiar@prplfoundation.org>