ci: push sbom to DependencyTrack
Push sboms to dependency track on protected references.

Since we want to have analysis running for all our platforms, we use the name of the branch/tag as the name of the project to be analysed.
For the version we choose the name of the platform. This results in an analysis for every branch/tag for every board.

Closes: PCF-1501

Signed-off-by: Matthias FRANCK <matthias.franck@softathome.com>