profiles: prpl: bump feed-prpl to latest HEAD
Bump feed-prpl to currently latest commit on the main branch.
This includes following fix:
prpl-configuration: rename cthulhu_usp.odl

When compiling with latest prpl-configuration package, we see:
 * check_data_file_clashes: Package prpl-configuration wants to install file /builds/prpl-foundation/prplos/prplos/build_dir/target-aarch64_cortex-a73+neon-vfpv4_musl/root-ipq95xx/etc/amx/cthulhu/extensions/cthulhu_usp.odl
	But that file is already provided by package  * cthulhu-plugin-usp
 * opkg_install_cmd: Cannot install package prpl-configuration.

The cthulhu_usp.odl is already provided in cthulhu-plugin-usp. Since both files need to be installed, let's rename this file so that they do not clash.

Signed-off-by: Matthias FRANCK <matthias.franck@softathome.com>