profiles: prpl: feed_wifi_core: pwhm with Multi-AP support.
This includes following fix:
Set multi_ap_profile in the wpa_supplicant config network section.
Since wld_wpaCtrl_events does not handle any internal ctx, nor executing any dm update
there was created a new handler for MultiAPInfo that will be called when profile/vlanId
are available in the message so that it can be set in the data model.

Freedom is not including the Multi-AP Profile subelement in its Association Request.
This is caused by missing implementation on Freedom’s WHM side to set multi_ap_profile
in the wpa_supplicant config.

Closes: PPM-3117.

Signed-off-by: Dmytro Puz <d.puz@inango-systems.com>