ci: cram: Update the post script crash check
ProcessFaults.LastUpgradeCount keeps the crash count, even after clearing the crashes. As the cram tests and the CDRouter tests use the same post script, it should be such a test that works in both cases.

cram tests have a simulated crash (setting LastUpgradeCount to 1). In CDRouter LastUpgradeCount should be 0. By checking the contents of ProcessFaults we know if there were any crashes with the added benefit of seeing what crashed.

The check of the ProcessFaults parameters was move to the ProcessFaults cram test

Signed-off-by: Toon Caeyers <toon.caeyers_ext@softathome.com>