PROVISORY: ci: cram: Remove missing IncrementalResult parameter from expected results
This parameter is missing according to the cram test but when performing the same check manually on a board, the parameter is there.

We are removing the parameter from the test in order for the test to pass while we investigate the issue.

Issue: PPW-400 - X_PRPL-COM_DownloadResult.IncrementalResult missing
Signed-off-by: Toon Caeyers <toon.caeyers_ext@softathome.com>