profiles: prpl: merge feed_wifi_swl and feed_wifi_core into new feed_prplmesh
For the time being we’ve accepted internal prplWare feed structure as
being used within SoftAtHome.

This feed structure doesn’t fit into the working group structure and
workflows within the prpl Foundation, so it was decided, that it should
be changed.

pwhm and prplMesh are now part of one single feed_prplmesh, so lets
adapt the feed configuration accordingly.

References: PPM-3113, PCF-1554
Signed-off-by: Petr Štetiar <petr.stetiar@prplfoundation.org>