ci: allow to bypass integrity check
When we want to test development branches of components, we need to point to the gitrepo and branch/hash/version we want to checkout of the component(s).
The CI has a check to see if the checksum of the corresponding package is correctly set in the package makefiles.

With this change, one can set the "skip_integrity_check" label on a MR.
In this case, the integrity check is not performed and we can simply download and use the development branch of the gitrepo.
This allows us to make MRs on the fly for testing purposes. Without needing to calculcate the checksum every time.

References: PPW-408

Signed-off-by: Matthias Franck <matthias.franck@softathome.com>