ci: testbed-device.py: fix VLAN configuration on OSPv2 and OSPv1 after DTS changes
Currently the CDRouter VLAN tests are failing with:

 FAIL: Could not establish IPv4 DHCP address on lan » no-lan-dhcp-address

This is happening as the UPDK 9.1.80 upstep changes the value of the
model variables in the device tree:

 * OSPv1: from `mxl,lightning mountainmxl,lgp` to `mxl,osp-tb341`
 * OSPv2: from `mxl,lightning mountainmxl,octopusmxl,octopus-641` to `mxl,osp-tb341-v2`

and thus the DUT VLAN configuration is unable to detect the board
properly and configure the VLAN on the the correct LAN port properly.

So lets adapt the board identification in the VLAN configuration to the
new device tree model variables.

Fixes: 45d2f1fcccc7 ("profiles: mxl_x86_osp_tb341_v2: update to UPDK 9.1.80")
Signed-off-by: Petr Štetiar <petr.stetiar@prplfoundation.org>