ci: mark tests which could fail as allowed to fail to maintain green CI status
Currently we're seeing a bit confusing workflow, where merge requests
with failed (red) CI pipeline are being merged. This is caused by the
fact, that sometimes we run additional tests on CI testing platforms
like Haze/Turris, which are not properly maintained due to limited
resources, so it takes a bit more time to fix those CI tests.

Sometimes we run test cases which are work in progress, like the prpl
Certification or pending fixes delivery like UPnP. It doesn't mean, that
there is necessarily something wrong with the contribution under the
test.

We're just making sure, that no new regressions are being introduced or
for example, when adding new tests, that they're generic enough and
works across the boards.

So lets improve the situation and mark such test cases as allowed to
fail, so once we use them in the pipeline for internal testing purposes,
their result doesn't influence the resulting pipeline status and thus we
would be merging contributions with possibly only OK (green) CI
pipeline.

Suggested-by: Frederik Van Bogaert <frederik.vanbogaert@prplfoundation.org>
Signed-off-by: Petr Štetiar <petr.stetiar@prplfoundation.org>