PROVISORY: ci: lcm: restart cthulhu before usp test
The LCM USP plugin doesn't support the restart of the USP broker.
Support of USP broker restart will be handled in LCM-835.

Currently, if the USP broker is restarted, the LCM USP functionnality won't
be usable and the tests will fail.
To avoid such a problem, this commit adds an explicit restart of
cthulhu before starting the LCM USP tests to workaround the missing
support of broker restart.

References: LCM-835
Signed-off-by: Nabil BIZID <nabil.bizid@softathome.com>