ci: define time limits for runtime testing job limits
Currently the default CI job timeout is set to 2 hours globally in
GitLab.com settings, which is quite a lot for most of the runtime
testing jobs.

Sometimes we get testing issues, which means that the testbed is
blocked unnecessarily for 2 hours.

To improve the situation, we define lower, job-specific timeouts:

 * cram to 30 min
 * CDRouter default 40 min, Wireless 15 min, UPnP 70 min, Certification 2h

This change ensures that jobs will fail faster and resources can be
freed up in a timely manner.

Signed-off-by: Petr Štetiar <petr.stetiar@prplfoundation.org>