scripts: gen_config.py: add support for replacing component with local component
This adaptation intends to supports CI on component level.
When the environment variable "LOCAL_COMPONENT" is defined in the environment the script will try to find this component in the
feeds directory.
When the component is found, the Makefile (and Config.in) for this package will be replaced with the newly generated one from baf
(pointing to the development branch of this component).

This way, one can directly make a build on the merge request of a component.

Signed-off-by: Matthias Franck <matthias.franck@softathome.com>