feeds.conf.default: bump packages feed to fix libxslt compilation
libxml2 was updated in OpenWrt Git tree with commit dec59db8fb1f
("libxml2: update to 2.13.6"), which fixed several CVEs.

Unfortunately this version bump included removal of some features, which
leads to build issues of libxslt:

  libxslt-1.1.37/xsltproc/xsltproc.c:733:39: error: assignment of read-only variable 'xmlParserMaxDepth'
   733 |                     xmlParserMaxDepth = value;

So lets fix it by backporting an upstream "fix", which removes that
deprecated functionality.

Fixes: dec59db8fb1f ("libxml2: update to 2.13.6")

Signed-off-by: Matthias Franck <matthias.franck@softathome.com>