scripts/mkits.sh: improve error handling for required parameters
Currently the script will exit with usage information and an error if a required
parameter is missing, but it actually doesn't provide any information about which
parameter is missing, so its quite tedious to find out which parameter is missing.

So lets fix it by adding a function to check for missing parameters and printing
a proper error message with the missing parameter(s):

  Error: Missing required parameters:
    CONFIG (-c): config name
  Usage: mkits.sh -A arch -C comp -a addr -e entry -v version -k kernel [-D name -n address -d dtb] -o its_file -c config_name
  ...

While at it, harmonize the shebang with other shell scripts and thus as
well allow usage of more advanced bash features.

Additionally fix the missing usage description of `-c` argument.

Upstream-Status: Submitted [https://github.com/openwrt/openwrt/pull/18524]
Signed-off-by: Petr Štetiar <ynezz@true.cz>
Signed-off-by: Petr Štetiar <petr.stetiar@prplfoundation.org>