pWHM: add FSM sync fix from sah-next
Due to some race condition between WHM's FSM and initialising a VAP, it's possible a hostapd BSS configuration gets configured with a zero MAC.
Mitigate this with an extra check when the initialisation is done.

This is supposed to be fixed in pwhm (sah-next): 2b3682cfa9899d724b68285001433fb0902f5ce6
Add this commit as a patch in order to test it.

Signed-off-by: Maarten De Decker <maarten.dedecker@mind.be>