profiles: prpl: feed-prplos: unbound-prpl: unbound-control: views and interfaces
It is not possible to reconfigure views and interfaces without restarting Unbound. Restarting is something time consuming and we want to avoid this as much as possible. (also we lose the cache and other information learned)

To make it possible to add/remove views and interfaces without restarting we extend UnboundControl with the following cmd arguments:

	* "interface <interface name>,<int position>": start listening on interface, pos is either 0 (insert at beginning) or another position to insert it at in the list of interfaces
	* "interface_remove <interface name>": stop listening on interface
	* "view <view name>,<view-first: yes or no>": add a view
	* "view_remove <view name>": remove a view

Reference: PCF-1057

Signed-off-by: Matthias Franck <matthias.franck@softathome.com>