target: don't use GENERIC_PLATFORM_DIR for generic other-files
Generic other-files directory path hardly needs to be overwritten an 99%
of the time the mainline generic one is usable.

Don't depend on GENERIC_PLATFORM_DIR variable (that can be redefined by
feeds) but introduce a dedicated define and set it to the default
location by default.

If needed feeds can still overwrite it.

Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>