fixup! TESTING DO NOT MERGE: ci: cdrouter: bulkdata testing
Currently we're having issues with CWMP tests and validation of
Device.BulkData.Protocols as it contains USPEventNotif protocol which is
USP only.

So lets try to workaround this by disabling cwmpDataValidation.

When this cwmpDataValidation is set to yes, CDRouter will perform data
validation on parameter values during CWMP profile testing. This means
the values of parameters will be checked to ensure they are
syntactically correct based on the specification found in the data
model.

Signed-off-by: Petr Štetiar <petr.stetiar@prplfoundation.org>