PROVISORY: profiles: cellular: enable lua-cjson package
In modemmanager we have a "sub"package called modemmanager-rpcd which depends on lua-cjson.
This package is not installed/enabled, but still the compile step of the global
modemmanager package is doing a global compilation, taking the lua-cjson into account.
This results in the lua-cjson being downloaded at compiletime and not in the "make download" phase of the build.

The osp boards build in offline mode, meaning internet connection is closed
after the "make download" step. As a result, the lua-cjson package cannot
be downloaded during the compile step.

As a workaround we set the lua-cjson package to "m",

Signed-off-by: Matthias Franck <matthias.franck@softathome.com>