scripts: gen_config: add support for local feeds
To point to a local feed, the 'method' key of the feed section can be set
to 'src-link' in a profile YAML file. However, the script first checks that
a revision is specified and aborts if not. This check is useless for a local feed.

Let's only check the 'revision' field for 'git-src' and 'git-src-full'
methods which are the only ones that supports a specific revision.

Signed-off-by: Cedric CHEDALEUX <cedric.chedaleux@orange.com>
(cherry picked from commit 49cce511e642cd9d70b96877a318d1c196df4345)