ci: cdrouter: start prplMesh in controller mode
During CDRouter tests, prplMesh is currently running in Agent mode (default).
This means cwmpd is unable to resolve the "Device.WiFi.DataElements." tree, because it is specific to the prplMesh controller.

Therefore, start prplMesh in controller mode before these tests are executed, so the "Device.WiFi.DataElements." datamodel is available.

Closes: PPM-3382

Signed-off-by: Maarten De Decker <maarten.dedecker@mind.be>