profiles: mxl_x86_osp_tb341_v2: fix kernel crash during radio re-enable scenario
Added a patch in iwlwav-driver-uci to Fix PPW-508 issue.

Rootcause: During the Wi-Fi radio disable/enable process, the kernel
process proc entry (/proc/net/mtlk/wlan3) was not properly removed.
As a result, when the radio band was “re-enabled”, the Wi-Fi driver
was unable to “recreate” the /proc/net/mtlk/wlan3 entry (return NULL),
leading to kernel panic issues.

Fix: Modify the _wv_ieee80211_op_remove_interface() function of MxL
 wi-fi driver to increase the retry duration for removing the proc entry.
The framework available is to try multiple times removing of the vaps
if the scan is running. This is basically increasing the retries if
removal of vap is failed due to scan.

Fixes PPW-508

Signed-off-by: Bhavesh Bharadiya <bnbharadiya@maxlinear.com>