ci: cram: fix wifi-scheduler.t script issue
Problem:
The wifi-scheduler.t script is sometimes failing with the error :
   $ delay_sec=$((60-$current_sec))
+  /bin/sh: 70: arithmetic expression: expecting EOF: "60-09" (no-eol)

Cause:
This issue happens when current_sec value is between 7 and 10
In shells like posix sh (but not bash), numbers with leading zeros are
treated as octal. Since for example 09 is not valid in octal, this
results in an error.

Solution:
Remove leading zero from current_sec to avoid this kind of wrong
interpretation.

Reference: PPM-2543

Signed-off-by: Houssem Dafdouf <houssem.dafdouf_ext@softathome.com>