profiles: qca_ipq95xx: qca-nss-ppe-ds: fix random qca-nss-ppe-ds compile failures in parallel builds
It was reported that qca-nss-ppe-ds fails to build when running make
-j33. This issue occurs because ppe_vp_public.h is not yet available
when qca-nss-ppe-ds is being built. Although kmod-qca-nss-ppe-vp is
listed in the DEPENDS of the qca-nss-ppe-ds Makefile, this does not
guarantee the required header is ready in time.

To ensure qca-nss-ppe-ds is built only after kmod-qca-nss-ppe-vp has
completed its InstallDev step, add kmod-qca-nss-ppe-vp to
PKG_BUILD_DEPENDS in the qca-nss-ppe-ds Makefile.

Fixes: PCF-1967

Signed-off-by: Marvin Lu <marvin.lu@wnc.com.tw>