ci: release: add Arcadyan Mozart bootstack assets
Arcadyan Mozart supports special recovery procedure and permits to
update the bootloader hence it might be handy to provide the bootloader
file in a release.

Add the bootloader assets to the release in a single TAR called
bootstack that include the Preloader, the GPT table and the
BL31+U-Boot binary.

References: PCF-2005
Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>
(cherry picked from commit 30830d1183179d5e15093993e168e9faba66dad7)