prpl.yml, cram tests: fix certificate related cram tests
020-security.t is self contained in terms of input and expected output.
Since the content of /etc/config/autocert directory in the system is no
longer empty after softhsm integration, adjust the test to save and
restore the system test before and after its execution.

Also change hash in feeds_prplos to integrate the change to softhsm
initialization to make it not as early in the boot process.

Signed-off-by: Piotr Nakraszewicz <piotr.nakraszewicz@consult.red>