PROVISORY: ci: cram: mxl25641-hdk-6: 110-prplmesh-dataelements: disable test checking prplmesh status
The test calls the script "/opt/prplmesh/scripts/prplmesh_utils.sh". The following error is seen when the script is called:
pgrep: pattern that searches for process name longer than 15 characters will result in zero matches
Try `pgrep -f' option to match against the complete command line.

So let's disable this test for now to be able to continue.

Signed-off-by: Matthias Franck <matthias.franck@softathome.com>