ci: cram: properly stop/start prplmesh during pwhm.t execution
Problem:
- During the pwhm.t cram test, the wld process gets somtimes stuck when
enabling or disabling VAPs. This only happens when the direct socket is
used.

Cause:
- When the pwhm.t test starts, we suspend the beerocks processes using
SIGSTOP.
- When pwhm raises an event through the USP socket, it expects an
acknowledgement from prplmesh (the client). Since prplmesh is suspended
it never replies.
- This concerns ony the direct socket case, which is point-to-point.
With ubus, events are broadcasted, so no pear acknowledgement is
required.
- The issue became visible after unfiltering custom events in amxb_usp.

Solution:
Stop prplmesh properly before starting the script so that it cleanly
disconnects from the USP backend.

Fixes: PPW-735

Signed-off-by: Houssem Dafdouf <houssem.dafdouf_ext@softathome.com>