ci: cram: wifi-scheduler.t: current_sec computation fails
Problem:
The script fails sometimes with the error :
bin/sh: 70: arithmetic expression: expecting primary: "60-" (no-eol)

Cause:
When '$(R date +%S)' is equal to zero, we unintentionally reset the
'current_sec' variable to an empty value because of 'sed 's/^0//'',
which was used to avoid octal interpretation.

Solution:
Calculate delay_sec differently so we don't depend on format an also
shell.

Closes: PPM-3507

Signed-off-by: Houssem Dafdouf <houssem.dafdouf_ext@softathome.com>