PROVISORY: ci: cdrouter: mxl25641-hdk-6: extend prplOS-Network-Test-Plan-Candidate-v1.1.0 timeout by 5 minutes
[ In commit 0395f03cc0e6 ("PROVISORY: ci: cdrouter: mxl25641-hdk-6:
  extend prplWare-Smoke-Test-Suite timeout by 5 minutes") we've increased
  test runtime by 5 minutes, likely due to the same reasons. ]

This provisional change extends the timeout from 90 to 95 minutes to prevent
CI failures while the root cause of the performance degradation is being
investigated.

The performance issue appears to be specific to the OSPv2 platform on the
latest-24.10 branch, which runs consistently slower than other platforms
(Freedom, Mozart) and has shown a recent increase beyond its historical
average.

This is a temporary workaround. The underlying performance regression still
needs to be identified and resolved.

References: PPW-908
Link: https://gitlab.com/prpl-foundation/prplos/prplos/-/jobs/11936157538#L199
Signed-off-by: Petr Štetiar <petr.stetiar@prplfoundation.org>